+++ src/components/Pills/Pill.tsx
@@ -137,6 +137,11 @@ export const Pill: FC<PillProps> = React.forwardRef(
             badgeProps={{ classNames: styles.badge }}
             iconProps={{ path: IconName.mdiClose }}
             {...closeButtonProps}
+            ariaLabel={
+              closeButtonProps?.ariaLabel ??
+              closeButtonProps?.getAriaLabel?.(label) ??
+              `Delete ${label || 'item'}`
+            }
             onClick={!mergedDisabled ? onClose : null}
             size={pillSizeToButtonSizeMap.get(size)}
             classNames={styles.closeButton}